Add one and zero trait
